Added debug information tracking
In this version, I introduced a global variable printDebugInfo, which controls whether debugging information should be printed or not. By default, it's set to true, so debug information will be printed. However, users can set it to false to disable printing of debug information before calling IsProgramDebugged(). This way, users don't need to pass any arguments to control the printing behavior.